+++ src/external-review/external-review-normalization.ts
@@ -1,7 +1,7 @@
-import { truncate } from "./utils";
-import { type LocalReviewSeverity } from "./local-review/types";
+import { truncate } from "../utils";
+import { type LocalReviewSeverity } from "../local-review/types";
 import { hasActionableReviewText, isActionableTopLevelReview } from "./external-review-signal-heuristics";
-import { type IssueComment, type PullRequestReview, type ReviewThread } from "./types";
+import { type IssueComment, type PullRequestReview, type ReviewThread } from "../types";
 
 export type ExternalReviewSignalSourceKind = "review_thread" | "top_level_review" | "issue_comment";
