+++ src/external-review/external-review-signal-heuristics.ts
